Update search api call
