+++ src/App.js
@@ -1,20 +1,14 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect, useState } from "react";
 import {
   Routes,
   Route,
-  createSearchParams,
   useSearchParams,
-  useNavigate,
 } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import "reactjs-popup/dist/index.css";
-import { fetchMovies } from "./data/moviesSlice";
-import {
-  ENDPOINT_SEARCH,
-  ENDPOINT_DISCOVER,
-  ENDPOINT,
-  API_KEY,
-} from "./constants";
+import { fetchMovies, fetchSingleMovie } from "./data/moviesSlice";
+import { ENDPOINT_SEARCH, ENDPOINT_DISCOVER } from "./constants";
 import Header from "./components/Header";
 import Movies from "./components/Movies";
 import Starred from "./components/Starred";
@@ -25,73 +19,49 @@ import "./app.scss";
 
 const App = () => {
   const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const { movies, fetchStatus } = useSelector((state) => state.movies);
+  const { movies } = useSelector((state) => state.movies);
   const [searchParams, setSearchParams] = useSearchParams();
   const searchQuery = searchParams.get("search");
-
   const [videoKey, setVideoKey] = useState();
   const [isOpen, setOpen] = useState(false);
 
   const closeModal = () => setOpen(false);
 
   const getSearchResults = (query) => {
-    if (query !== "") {
+    if (query) {
       dispatch(fetchMovies(`${ENDPOINT_SEARCH}?query=${query}`));
-      setSearchParams(createSearchParams({ search: query }));
     } else {
       dispatch(fetchMovies(ENDPOINT_DISCOVER));
-      setSearchParams();
     }
   };
 
-  const searchMovies = (query) => {
-    navigate("/");
-    getSearchResults(query);
-  };
-
   const getMovies = () => {
-    if (fetchStatus === "loading") return;
     if (searchQuery) {
-      dispatch(fetchMovies(`${ENDPOINT_SEARCH}?query=${searchQuery}&page=1`));
+      dispatch(fetchMovies(`${ENDPOINT_SEARCH}?query=${searchQuery}`));
     } else {
-      dispatch(fetchMovies(`${ENDPOINT_DISCOVER}?page=1`));
+      dispatch(fetchMovies(`${ENDPOINT_DISCOVER}`));
     }
   };
 
   const viewTrailer = (movie) => {
     getMovie(movie.id);
-    if (!videoKey) setOpen(true);
     setOpen(true);
   };
 
   const getMovie = async (id) => {
-    const URL = `${ENDPOINT}/movie/${id}?api_key=${API_KEY}&append_to_response=videos`;
-
     setVideoKey(null);
-    try {
-      const videoData = await fetch(URL).then((response) => response.json());
-
-      if (videoData.videos && videoData.videos.results.length) {
-        const trailer = videoData.videos.results.find(
-          (vid) => vid.type === "Trailer"
-        );
-        setVideoKey(trailer ? trailer.key : videoData.videos.results[0].key);
-      }
-    } catch (error) {
-      console.error("Error fetching movie data:", error);
-    }
+    const res = await fetchSingleMovie(id);
+    if (res) setVideoKey(res);
   };
 
   useEffect(() => {
     getMovies();
-    //eslint-disable-next-line
   }, []);
 
   return (
     <div className="App">
       <Header
-        searchMovies={searchMovies}
+        searchMovies={getSearchResults}
         searchParams={searchParams}
         setSearchParams={setSearchParams}
       />
