+++ src/constants.js
@@ -1,13 +1,6 @@
 export const API_KEY = '8cac6dec66e09ab439c081b251304443'
-//the Endpoint has since been updated, it now requires Bearer Token to be passede in the request header
 export const TOKEN = 'eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIwMTUzZGY3Njc0OTE1NDFhODdhYWU1ZDg5MjJmNWQ1NSIsInN1YiI6IjY2NDY3YjA0ZGVmMjcwNTg5Y2E5MGE1NiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.y71quUtq5frJzB8_ojpiStM3z5NevDbF244pVWEN30c'
 export const ENDPOINT = 'https://api.themoviedb.org/3'
 export const ENDPOINT_DISCOVER = ENDPOINT+'/discover/movie'
 export const ENDPOINT_SEARCH = ENDPOINT+'/search/movie'
 export const ENDPOINT_MOVIE = ENDPOINT+'/movie/507086?api_key='+API_KEY+'&append_to_response=videos'
-
-//0153df767491541a87aae5d8922f5d55
-
-/*
-eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIwMTUzZGY3Njc0OTE1NDFhODdhYWU1ZDg5MjJmNWQ1NSIsInN1YiI6IjY2NDY3YjA0ZGVmMjcwNTg5Y2E5MGE1NiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.y71quUtq5frJzB8_ojpiStM3z5NevDbF244pVWEN30c
-*/
